QMAPS-2358 history-suggest wip
