+++ src/components/ui/Suggest.jsx
@@ -2,10 +2,12 @@ import React, { useEffect, useState, useCallback, useRef } from 'react';
 import ReactDOM from 'react-dom';
 import debounce from 'lodash.debounce';
 import { bool, string, func, object } from 'prop-types';
-import { useDevice, useI18n } from 'src/hooks';
+import { useConfig, useDevice, useI18n } from 'src/hooks';
 import SuggestsDropdown from 'src/components/ui/SuggestsDropdown';
 import { fetchSuggests, getInputValue, modifyList } from 'src/libs/suggest';
 import { UserFeedbackYesNo } from './index';
+import { getHistoryPrompt, setHistoryPrompt, setHistoryEnabled } from 'src/adapters/search_history';
+import { Box, Button, Heading, Stack } from '@qwant/qwant-ponents';
 
 const SUGGEST_DEBOUNCE_WAIT = 100;
 
@@ -45,18 +47,97 @@ const Suggest = ({
   withFeedback,
   hide,
 }) => {
+  const searchHistoryConfig = useConfig('searchHistory');
+
   const [items, setItems] = useState([]);
   const [isOpen, setIsOpen] = useState(false);
   const [highlighted, setHighlighted] = useState(null);
   const [hasFocus, setHasFocus] = useState(false);
+  const [answer, setAnswer] = useState(null);
+  const historyPromptVisible =
+    isOpen &&
+    searchHistoryConfig?.enabled &&
+    value === '' &&
+    (getHistoryPrompt() === null || answer !== null);
   const dropdownVisible = hasFocus && isOpen && outputNode;
   const { isMobile } = useDevice();
   const { _ } = useI18n();
   const dropDownContent = useRef();
 
   const close = () => {
-    setIsOpen(false);
+    if (!historyPromptVisible) {
+      setIsOpen(false);
+    }
     setItems([]);
+  };
+
+  const historyPrompt = () => {
+    if (answer === null) {
+      return (
+        <Box m="l">
+          <Heading as="h6">{_('History is available on Qwant Maps', 'history')}</Heading>
+          <Stack>
+            <Box>
+              {_(
+                'Convenient and completely private, the history will only be visible to you on this device 🙈.',
+                'history'
+              )}{' '}
+              <a href="">{_('Read more', 'history')}</a>
+            </Box>
+            <Box mt="l">
+              <Button
+                variant="secondary"
+                onClick={() => {
+                  setHistoryPrompt(true);
+                  setHistoryEnabled(false);
+                  setAnswer(false);
+                }}
+              >
+                {_('No thanks', 'history')}
+              </Button>
+              <Button
+                ml="l"
+                onClick={() => {
+                  setHistoryPrompt(true);
+                  setHistoryEnabled(true);
+                  setAnswer(true);
+                }}
+              >
+                {_('Enable history', 'history')}
+              </Button>
+            </Box>
+          </Stack>
+        </Box>
+      );
+    } else if (answer === true) {
+      return (
+        <Box m="l">
+          <Heading as="h6">{_('Well done, the history is activated!', 'history')}</Heading>
+          <Stack>
+            <Box>
+              {_(
+                'You can find and manage your complete history at any time in the menu',
+                'history'
+              )}
+            </Box>
+          </Stack>
+        </Box>
+      );
+    } else if (answer === false) {
+      return (
+        <Box m="l">
+          <Heading as="h6">{_('No worries, history is disabled', 'history')}</Heading>
+          <Stack>
+            <Box>
+              {_(
+                'You can change your mind at any time and manage the activation of the history in the menu.',
+                'history'
+              )}
+            </Box>
+          </Stack>
+        </Box>
+      );
+    }
   };
 
   useEffect(() => {
@@ -178,16 +259,19 @@ const Suggest = ({
         },
         highlightedValue: highlighted ? getInputValue(highlighted) : null,
       })}
-      {dropdownVisible &&
+      {(dropdownVisible || historyPromptVisible) &&
         ReactDOM.createPortal(
           <div ref={dropDownContent}>
-            <SuggestsDropdown
-              className={className}
-              suggestItems={items}
-              highlighted={highlighted}
-              onSelect={selectItem}
-              value={value}
-            />
+            {dropdownVisible && !historyPromptVisible && (
+              <SuggestsDropdown
+                className={className}
+                suggestItems={items}
+                highlighted={highlighted}
+                onSelect={selectItem}
+                value={value}
+              />
+            )}
+            {historyPromptVisible && historyPrompt()}
             {withFeedback && value && items.length > 0 && !items[0].errorLabel && (
               <UserFeedbackYesNo
                 questionId="suggest"
@@ -198,6 +282,7 @@ const Suggest = ({
           </div>,
           outputNode
         )}
+      {}
     </>
   );
 };
